+++ src/kafunk/Utility/AsyncSeq.fs
@@ -196,9 +196,11 @@ module AsyncSeq =
           match rem with
           | Some rem -> async.Return rem
           | None -> Async.StartChildAsTask (ie.MoveNext())
-        let t = Stopwatch.GetTimestamp()
-        let! time = Async.StartChildAsTask (Async.Sleep (max 0 rt))
-        let! moveOr = Async.chooseTasks move time
+        let t = Stopwatch.GetTimestamp()        
+        let time = Task.Delay (max 0 rt)
+        let! moveOr = Async.chooseTasksUnit move time
+//        let! time = Async.StartChildAsTask (Async.Sleep (max 0 rt))
+//        let! moveOr = Async.chooseTasks move time
         let delta = int ((Stopwatch.GetTimestamp() - t) * 1000L / Stopwatch.Frequency)
         match moveOr with
         | Choice1Of2 (None, _) -> 
@@ -246,3 +248,11 @@ module AsyncSeq =
       finally 
          // Cancel on early exit 
          cts.Cancel() }
+
+  let mergeChoice3 (s1:AsyncSeq<'a>) (s2:AsyncSeq<'b>) (s3:AsyncSeq<'c>) : AsyncSeq<Choice<'a, 'b, 'c>> =
+    AsyncSeq.mergeAll 
+      [ s1 |> AsyncSeq.map Choice1Of3 ; s2 |> AsyncSeq.map Choice2Of3 ; s3 |> AsyncSeq.map Choice3Of3 ]
+
+  let mergeChoice4 (s1:AsyncSeq<'a>) (s2:AsyncSeq<'b>) (s3:AsyncSeq<'c>) (s4:AsyncSeq<'d>) : AsyncSeq<Choice<'a, 'b, 'c, 'd>> =
+    AsyncSeq.mergeAll 
+      [ s1 |> AsyncSeq.map Choice1Of4 ; s2 |> AsyncSeq.map Choice2Of4 ; s3 |> AsyncSeq.map Choice3Of4 ; s4 |> AsyncSeq.map Choice4Of4 ]