+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 0.1.11 - 17.11.2017
+* BREAKING: Offset commit queue is moved to module `PeriodicCommitQueue`. See: https://jet.github.io/kafunk/consuming.html#Periodic-Offset-Commit.
+* Producer performance improvements.
+
 ### 0.1.10 - 15.11.2017
 * Fix another scenario referenced in 0.1.9 producer batching fix.